+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 ## 0.9a
 added:
-- AES-256, ChaCha20 encryption algorithms
+- AES-256, ChaCha20, ChaCha20-Poly1305 encryption algorithms
 
 fixed:
 - abort at build service process if token exists and authorization succeed